Split entr init into three modules. Have not yet fixed notebooks
